Deleted the TestScene, as it wasn't used anymore.
It was used to test the combat mechanics of the agent, as well as their movement. But it's no longer needed. Besides, it would cause errors to just simply spawn the agents on the scene, as the codebase was changed so that certain initialization methods are necessary before spawning the agents.