Added WebGL build support
On a WebGL build, the game window gets automatically resized whenever the browser window's dimensions are changed. This is thanks to "(Unity) Better Minimal WebGL Template" by Sean:

The credits to Sean are given in DocumentationMainPage.md as well, which can be found in the Assets/Scripts directory.

On a standalone Windows build, there is no change. On a WebGL build, the mouse cursor is locked if during combat so that the player can rotate their character without any problems. When the cursor is needed, it is unlocked and it becomes visible.